Use notes to track reviewid/lasthash.
This change is primarily to avoid RB email spam when RBs are "updated" with
diffs that have not actually changed. (E.g. if a user invokes review-branch
several times, but the commits earlier in the branch have not changed.)

By tracking the last tree hash we sent to RB for each commit, we can avoid
sending it a no-op update.

Tracking this tree id in the commit message (similar to the "RB=<rb-id>"
convention) would be annoying, and turns out git notes is a surprisingly good
place to put it.

Note that for notes to follow commits as they're rebased, the refs need to be
enabled in the config, which we do automatically for our special reviewid and
reviewlasttree notes.